+++ src/Components/EditAdditionalRepositoriesTable/index.js
@@ -1,0 +1,2 @@
+import EditAdditionalRepositoriesTable from './EditAdditionalRepositoriesTable';
+export { EditAdditionalRepositoriesTable as default };
